Moved getxmp() into JpegImageFile
