--- Tests/test_image_getxmp.py
@@ -1,9 +1,0 @@
-from PIL import Image
-
-
-def test_getxmp():
-    with Image.open("Tests/images/xmp_test.jpg") as im:
-        xmp = im.getxmp()
-
-        assert isinstance(xmp, dict)
-        assert xmp["Description"]["Version"] == "10.4"
